Use size_t instead of int on printer

+++ includes/printer.h
@@ -6,7 +6,7 @@
 /*   By: jleem <jleem@student.42seoul.kr>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/19 21:19:35 by jleem             #+#    #+#             */
-/*   Updated: 2021/05/27 15:49:24 by jleem            ###   ########.fr       */
+/*   Updated: 2021/05/27 15:52:42 by jleem            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,14 @@
 # define PRINTER_H
 
 # include <stdarg.h>
+# include <stddef.h>
 
 typedef struct	s_printer
 {
 	char const	*fmt;
-	int			fmt_idx;
-	int			fmt_len;
-	int			nchar;
+	size_t		fmt_idx;
+	size_t		fmt_len;
+	size_t		nchar;
 	va_list		*ap;
 }				t_printer;
